requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
gunicorn==22.0.0
python-dotenv==1.0.1
alembic==1.18.3
blinker==1.9.0
click==8.3.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
pytest==8.3.4
ruff==0.9.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5